Tidying up signature authentication - Renamed HttpMessageSigningContext to AspNetCoreMessageSigningContext - Moved types to separate files - Removed unused usings - Added todo item for selecting different verification signature providers
