+++ Letterbook.Api.Authentication.HttpSignature/HttpSignatureAuthenticationDefaults.cs
@@ -1,0 +1,6 @@
+namespace Letterbook.Api.Authentication.HttpSignature;
+
+public static class HttpSignatureAuthenticationDefaults
+{
+	public static readonly string Scheme = "HttpSignature";
+}
