+++ Letterbook.Api.Authentication.HttpSignature/HttpSignatureAuthenticationOptions.cs
@@ -1,0 +1,7 @@
+using Microsoft.AspNetCore.Authentication;
+
+namespace Letterbook.Api.Authentication.HttpSignature;
+
+public class HttpSignatureAuthenticationOptions : AuthenticationSchemeOptions
+{
+}
