+++ Letterbook.Api.Authentication.HttpSignature/IFederatedActorHttpSignatureVerifier.cs
@@ -1,0 +1,10 @@
+using Microsoft.AspNetCore.Http;
+
+namespace Letterbook.Api.Authentication.HttpSignature;
+
+public interface IFederatedActorHttpSignatureVerifier
+{
+	IAsyncEnumerable<Uri> VerifyAsync(
+		HttpContext context,
+		CancellationToken cancellationToken);
+}
